add: 列車走行位置を定期的に更新するように実装
