+++ src/app.ts
@@ -2,12 +2,12 @@ import { serve } from '@hono/node-server';
 import { Hono } from 'hono';
 import { logger } from 'hono/logger';
 import { app as indexPage } from './pages/index';
-import { app as trainMapPage } from './pages/TrainMap';
-import { app as locationApi, TrainState } from './api/location';
+import { app as trainMapPage } from './pages/trainMap';
+import { TrainState, trainLocation } from './train-controle/location';
 
 declare module 'hono' {
   interface ContextVariableMap {
-    state: { trainState: TrainState }
+    state: TrainState
   }
 }
 
@@ -15,7 +15,11 @@ const app = new Hono();
 
 app.use(logger());
 
-const state = { trainState: { t0: { mileage: 0.0 } } };
+const state = { t0: { mileage: 0.0 } };
+
+setInterval(() => {
+  trainLocation(state);
+}, 1000);
 
 app.use(async (c, next) => {
   c.set("state", state);
@@ -25,7 +29,6 @@ app.use(async (c, next) => {
 // pages
 app.route('/', indexPage);
 app.route('/trainMap', trainMapPage);
-app.route('/api/location', locationApi);
 
 const port = 3000;
 console.log(`Server is running on port ${port}`);
